Some small S2 improvements

+++ geometry/s1angle.cc
@@ -31,7 +31,7 @@ void S1Angle::Normalize() {
 ostream& operator<<(ostream& os, S1Angle const& a) {
   double degrees = a.degrees();
   char buffer[13];
-  int sz = snprintf(buffer, sizeof(buffer), "%.7f", degrees);
+  size_t sz = snprintf(buffer, sizeof(buffer), "%.7f", degrees);
   if (sz >= 0 && sz < sizeof(buffer)) {
     return os << buffer;
   } else {
